refactor(core): cleanup `egui` implementation
